HBX-1680: Move class 'org.hibernate.tool.hbm2x.MetaAttributeHelper' to 'org.hibernate.tool.internal.export.pojo.MetaAttributeHelper'
